requirements.txt
fastapi
uvicorn
pyserial
tk
pytest
